+++ public/index.html
@@ -9,14 +9,11 @@
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <link href="/your-path-to-fontawesome/css/fontawesome.css" rel="stylesheet"><link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-
     <title>Freelances Lyonnais</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
     <script src="https://kit.fontawesome.com/c49536602b.js" crossorigin="anonymous"></script>
-
-
   </body>
 </html>
