+++ src/components/Detail/Carousel.js
@@ -8,7 +8,7 @@ import KeyboardArrowLeft from '@material-ui/icons/KeyboardArrowLeft';
 import KeyboardArrowRight from '@material-ui/icons/KeyboardArrowRight';
 import SwipeableViews from 'react-swipeable-views';
 import { autoPlay } from 'react-swipeable-views-utils';
-import avatar from '../../images/reficon.png';
+import avatar from '../../images/logo2.png';
 
 const AutoPlaySwipeableViews = autoPlay(SwipeableViews);
