+++ src/index.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
-import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
   <React.StrictMode>
@@ -10,5 +9,3 @@ ReactDOM.render(
   document.getElementById('root')
 
 );
-
-serviceWorker.unregister();
