+++ src/pages/generic page/Home.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import logo from './logo.png';
+import logo from '../../images/logo.png';
 import { Helmet } from 'react-helmet';
 import { Link } from 'react-router-dom';
 import Button from '@material-ui/core/Button';
